+++ addons/ApplicationInsights/Tweek.Addons.ApplicationInsights/ApplicationInsightAddon.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Microsoft.ApplicationInsights.Channel;
 using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
@@ -9,24 +8,8 @@ using Tweek.ApiService.Addons;
 
 namespace Tweek.Addons.ApplicationInsights
 {
-    public class ApiClientTelemetryInitializer : ITelemetryInitializer
-    {
-        IHttpContextAccessor httpContextAccessor;
 
-        public ApiClientTelemetryInitializer(IHttpContextAccessor httpContextAccessor)
-        {
-            this.httpContextAccessor = httpContextAccessor;
-        }
-        public void Initialize(ITelemetry telemetry)
-        {
-            if (httpContextAccessor.HttpContext == null) return;
-            if (httpContextAccessor.HttpContext.Request.Headers.TryGetValue("x-api-client", out var values)){
-                telemetry.Context.Properties["x-api-client"] = values.ToString();
-            }
-        }
-    }
-
-    public class ApplicationInsightsAddon : ITweekAddon
+  public class ApplicationInsightsAddon : ITweekAddon
     {
         public void Use(IApplicationBuilder builder, IConfiguration configuration)
         {
